\name{Heatmap}
\alias{Heatmap}
\alias{Heatmap.default}
\title{Heatmap}
\description{
function for single environment analysis; outputs heatmap
}
\usage{
Heatmap(R, genAs = "fixed", row, column, respvar, model, env)
}
\arguments{
  \item{R}{data with residuals}
  \item{genAs}{a}
  \item{row}{variable name corresponding to field row}
  \item{column}{variable name corresponding to field column}
  \item{respvar}{vector of response variables}
  \item{model}{experimental design}
  \item{env}{variable name corresponding to environment column}
}
\author{Paul Eilers, Gerrit Gort, Sabine Schnabel, Alaine A. Gulles, Rose Imee Zhella Morantte}
\keyword{graph}
